+++ core/src/commonMain/kotlin/com/deezer/caupain/VersionCatalogParser.kt
@@ -1,8 +1,10 @@
 package com.deezer.caupain
 
+import com.deezer.caupain.model.Ignores
 import com.deezer.caupain.model.versionCatalog.VersionCatalog
 import com.deezer.caupain.serialization.DefaultToml
 import com.deezer.caupain.serialization.decodeFromPath
+import com.deezer.caupain.toml.IgnoreParser
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
@@ -13,7 +15,7 @@ import okio.Path
 import okio.SYSTEM
 
 internal interface VersionCatalogParser {
-    suspend fun parseDependencyInfo(): VersionCatalog
+    suspend fun parseDependencyInfo(): VersionCatalogParseResult
 }
 
 internal class DefaultVersionCatalogParser(
@@ -22,10 +24,22 @@ internal class DefaultVersionCatalogParser(
     private val fileSystem: FileSystem = FileSystem.SYSTEM,
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : VersionCatalogParser {
-    override suspend fun parseDependencyInfo(): VersionCatalog = withContext(ioDispatcher) {
-        toml.decodeFromPath(
-            path = versionCatalogPath,
-            fileSystem = fileSystem
-        )
-    }
+
+    private val ignoreParser = IgnoreParser(fileSystem, versionCatalogPath, ioDispatcher)
+
+    override suspend fun parseDependencyInfo(): VersionCatalogParseResult =
+        withContext(ioDispatcher) {
+            VersionCatalogParseResult(
+                versionCatalog = toml.decodeFromPath<VersionCatalog>(
+                    path = versionCatalogPath,
+                    fileSystem = fileSystem
+                ),
+                ignores = ignoreParser.computeIgnores()
+            )
+        }
 }
+
+internal data class VersionCatalogParseResult(
+    val versionCatalog: VersionCatalog,
+    val ignores: Ignores
+)